requirements.txt
# Core dependencies
aiohttp>=3.9.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
aiolimiter>=1.1.0

# Optional fallback
yfinance>=0.2.40

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-socket>=0.7.0
